+++ src/request-handler2.lisp
@@ -400,11 +400,12 @@ association list. This function is normally called by
                 ;; TODO: replace return-code with something else
                 (if (member weblocks.response:*code*
                             weblocks.variables:*approved-return-codes*)
-                    (progn 
+                    (let ((content (get-output-stream-string weblocks::*weblocks-output-stream*)))
                       (unless (weblocks.request:ajax-request-p)
-                        (weblocks.session::set-value 'last-request-uri
-                                                     (weblocks::all-tokens weblocks::*uri-tokens*)))
-                      (get-output-stream-string weblocks::*weblocks-output-stream*))
+                        (weblocks.session:set-value 'last-request-uri
+                                                    (weblocks::all-tokens weblocks::*uri-tokens*)))
+                      ;; Return rendered content as a response on request.
+                      content)
                     (weblocks::handle-http-error app weblocks.response:*code*)))))))
 
     ;; Restart